WIP break up reducers
