+++ packages/math-input/src/store/index.ts
@@ -1,274 +1,20 @@
 import * as Redux from "redux";
 
-import {tabletCutoffPx} from "../components/common-style";
-import {computeLayoutParameters} from "../components/compute-layout-parameters";
-import ExpressionKeypad from "../components/expression-keypad";
-import FractionKeypad from "../components/fraction-keypad";
 import GestureManager from "../components/gesture-manager";
-import * as CursorContexts from "../components/input/cursor-contexts";
-import VelocityTracker from "../components/velocity-tracker";
-import {
-    DeviceOrientations,
-    DeviceTypes,
-    EchoAnimationTypes,
-    KeyTypes,
-    KeypadTypes,
-    LayoutModes,
-} from "../consts";
-import KeyConfigs from "../data/key-configs";
 import Keys from "../data/keys";
 
-import type {Cursor, KeyHandler} from "../types";
-import type {
-    InputState,
-    KeypadState,
-    PagerState,
-    GestureState,
-    EchoState,
-    LayoutState,
-} from "./types";
+import echoReducer from "./echo-reducer";
+import inputReducer from "./input-reducer";
+import keypadReducer from "./keypad-reducer";
+import layoutReducer from "./layout-reducer";
+import pagerReducer from "./pager-reducer";
+import {defaultKeypadType, keypadForType} from "./shared";
 
-const keypadForType = {
-    [KeypadTypes.FRACTION]: FractionKeypad,
-    [KeypadTypes.EXPRESSION]: ExpressionKeypad,
-} as const;
+import type {GestureState} from "./types";
 
 export const createStore = () => {
-    const initialInputState: {
-        keyHandler: KeyHandler | null;
-        cursor: Cursor;
-    } = {
-        keyHandler: null,
-        cursor: {
-            context: CursorContexts.NONE,
-        },
-    };
-
-    const inputReducer = function (
-        state = initialInputState,
-        action: any,
-    ): InputState {
-        switch (action.type) {
-            case "SetKeyHandler":
-                return {
-                    ...state,
-                    keyHandler: action.keyHandler,
-                };
-
-            case "PressKey":
-                const keyConfig = KeyConfigs[action.key];
-                if (keyConfig.type !== KeyTypes.KEYPAD_NAVIGATION) {
-                    // This is probably an anti-pattern but it works for the
-                    // case where we don't actually control the state but we
-                    // still want to communicate with the other object
-                    return {
-                        ...state,
-                        cursor: state.keyHandler?.(keyConfig.id),
-                    };
-                }
-
-                // TODO(kevinb) get state from MathQuill and store it?
-                return state;
-
-            case "SetCursor":
-                return {
-                    ...state,
-                    cursor: action.cursor,
-                };
-
-            default:
-                return state;
-        }
-    };
-
-    const defaultKeypadType = KeypadTypes.EXPRESSION;
-
-    const initialKeypadState = {
-        extraKeys: ["x", "y", Keys.THETA, Keys.PI],
-        keypadType: defaultKeypadType,
-        active: false,
-    } as const;
-
-    const keypadReducer = function (
-        state = initialKeypadState,
-        action: {
-            type: string;
-        },
-    ): KeypadState {
-        switch (action.type) {
-            case "DismissKeypad":
-                return {
-                    ...state,
-                    active: false,
-                };
-
-            case "ActivateKeypad":
-                return {
-                    ...state,
-                    active: true,
-                };
-
-            case "ConfigureKeypad":
-                return {
-                    ...state,
-                    // Default `extraKeys` to the empty array.
-                    extraKeys: [],
-                    // @ts-expect-error [FEI-5003] - TS2339 - Property 'configuration' does not exist on type '{ type: string; }'.
-                    ...action.configuration,
-                };
-
-            case "PressKey":
-                // @ts-expect-error [FEI-5003] - TS2339 - Property 'key' does not exist on type '{ type: string; }'.
-                const keyConfig = KeyConfigs[action.key];
-                // NOTE(charlie): Our keypad system operates by triggering key
-                // presses with key IDs in a dumb manner, such that the keys
-                // don't know what they can do--instead, the store is
-                // responsible for interpreting key presses and triggering the
-                // right actions when they occur. Hence, we figure off a
-                // dismissal here rather than dispatching a dismiss action in
-                // the first place.
-                if (keyConfig.id === Keys.DISMISS) {
-                    return keypadReducer(state, {type: "DismissKeypad"});
-                }
-                return state;
-
-            default:
-                return state;
-        }
-    };
-
-    // We default to the right-most page. This is done so-as to enforce a
-    // consistent orientation between the view pager layout and the flattened
-    // layout, where our default page appears on the far right.
-    const getDefaultPage = (numPages) => numPages - 1;
-
-    const initialPagerState = {
-        animateToPosition: false,
-        currentPage: getDefaultPage(keypadForType[defaultKeypadType].numPages),
-        // The cumulative differential in the horizontal direction for the
-        // current swipe.
-        dx: 0,
-        numPages: keypadForType[defaultKeypadType].numPages,
-        pageWidthPx: 0,
-        velocityTracker: new VelocityTracker(),
-    } as const;
-
-    const pagerReducer = function (
-        state = initialPagerState,
-        action: {
-            type: string;
-        },
-    ): PagerState {
-        switch (action.type) {
-            case "ConfigureKeypad":
-                // @ts-expect-error [FEI-5003] - TS2339 - Property 'configuration' does not exist on type '{ type: string; }'.
-                const {keypadType} = action.configuration;
-                const {numPages} = keypadForType[keypadType];
-                return {
-                    ...state,
-                    numPages,
-                    animateToPosition: false,
-                    currentPage: getDefaultPage(numPages),
-                    dx: 0,
-                };
-
-            case "SetPageSize":
-                return {
-                    ...state,
-                    // @ts-expect-error [FEI-5003] - TS2339 - Property 'pageWidthPx' does not exist on type '{ type: string; }'.
-                    pageWidthPx: action.pageWidthPx,
-                };
-
-            case "PressKey":
-                // @ts-expect-error [FEI-5003] - TS2339 - Property 'key' does not exist on type '{ type: string; }'.
-                const keyConfig = KeyConfigs[action.key];
-
-                // Reset the keypad page if the user performs a math operation.
-                if (
-                    keyConfig.type === KeyTypes.VALUE ||
-                    keyConfig.type === KeyTypes.OPERATOR
-                ) {
-                    return pagerReducer(state, {type: "ResetKeypadPage"});
-                }
-                return state;
-
-            case "ResetKeypadPage":
-                return {
-                    ...state,
-                    animateToPosition: true,
-                    // We start at the right-most page.
-                    currentPage: getDefaultPage(state.numPages),
-                    dx: 0,
-                };
-
-            case "PageKeypadRight":
-                const nextPage = Math.min(
-                    state.currentPage + 1,
-                    state.numPages - 1,
-                );
-                return {
-                    ...state,
-                    animateToPosition: true,
-                    currentPage: nextPage,
-                    dx: 0,
-                };
-
-            case "PageKeypadLeft":
-                const prevPage = Math.max(state.currentPage - 1, 0);
-                return {
-                    ...state,
-                    animateToPosition: true,
-                    currentPage: prevPage,
-                    dx: 0,
-                };
-
-            case "OnSwipeChange":
-                // @ts-expect-error [FEI-5003] - TS2339 - Property 'dx' does not exist on type '{ type: string; }'.
-                state.velocityTracker.push(action.dx);
-
-                return {
-                    ...state,
-                    animateToPosition: false,
-                    // @ts-expect-error [FEI-5003] - TS2339 - Property 'dx' does not exist on type '{ type: string; }'.
-                    dx: action.dx,
-                };
-
-            case "OnSwipeEnd":
-                const {pageWidthPx, velocityTracker} = state;
-                // @ts-expect-error [FEI-5003] - TS2339 - Property 'dx' does not exist on type '{ type: string; }'.
-                const {dx} = action;
-                const velocity = velocityTracker.getVelocity();
-
-                // NOTE(charlie): These will need refinement. The velocity comes
-                // from Framer.
-                const minFlingVelocity = 0.1;
-                const minFlingDistance = 10;
-
-                const shouldPageRight =
-                    dx < -pageWidthPx / 2 ||
-                    (velocity < -minFlingVelocity && dx < -minFlingDistance);
-
-                const shouldPageLeft =
-                    dx > pageWidthPx / 2 ||
-                    (velocity > minFlingVelocity && dx > minFlingDistance);
-
-                if (shouldPageRight) {
-                    return pagerReducer(state, {type: "PageKeypadRight"});
-                } else if (shouldPageLeft) {
-                    return pagerReducer(state, {type: "PageKeypadLeft"});
-                }
-
-                return {
-                    ...state,
-                    animateToPosition: true,
-                    dx: 0,
-                };
-
-            default:
-                return state;
-        }
-    };
-
+    // TODO(matthewc): gestureReducer can't be moved from this file
+    // because it depends on `store` being in scope (see note below)
     const createGestureManager = (swipeEnabled: boolean) => {
         return new GestureManager(
             {
@@ -351,182 +97,6 @@ export const createStore = () => {
                     popover: null,
                     focus: null,
                     gestureManager: createGestureManager(swipeEnabled),
-                };
-
-            default:
-                return state;
-        }
-    };
-
-    // Used to generate unique animation IDs for the echo animations. The actual
-    // values are irrelevant as long as they are unique.
-    let _lastAnimationId = 0;
-
-    const initialEchoState = {
-        echoes: [],
-    } as const;
-
-    const echoReducer = function (state = initialEchoState, action): EchoState {
-        switch (action.type) {
-            case "PressKey":
-                const keyConfig = KeyConfigs[action.key];
-
-                // Add in the echo animation if the user performs a math
-                // operation.
-                if (
-                    keyConfig.type === KeyTypes.VALUE ||
-                    keyConfig.type === KeyTypes.OPERATOR
-                ) {
-                    return {
-                        ...state,
-                        echoes: [
-                            ...state.echoes,
-                            {
-                                animationId: "" + _lastAnimationId++,
-                                animationType: action.inPopover
-                                    ? EchoAnimationTypes.LONG_FADE_ONLY
-                                    : EchoAnimationTypes.FADE_ONLY,
-                                borders: action.borders,
-                                id: keyConfig.id,
-                                initialBounds: action.initialBounds,
-                            },
-                        ],
-                    };
-                }
-                return state;
-
-            case "RemoveEcho":
-                const remainingEchoes = state.echoes.filter((echo) => {
-                    // @ts-expect-error [FEI-5003] - TS2339 - Property 'animationId' does not exist on type 'never'.
-                    return echo.animationId !== action.animationId;
-                });
-                return {
-                    ...state,
-                    echoes: remainingEchoes,
-                };
-
-            default:
-                return state;
-        }
-    };
-
-    const initialLayoutState = {
-        gridDimensions: {
-            numRows: keypadForType[defaultKeypadType].rows,
-            numColumns: keypadForType[defaultKeypadType].columns,
-            numMaxVisibleRows: keypadForType[defaultKeypadType].maxVisibleRows,
-            numPages: keypadForType[defaultKeypadType].numPages,
-        },
-        buttonDimensions: {
-            widthPx: 48,
-            heightPx: 48,
-        },
-        pageDimensions: {
-            pageWidthPx: 0,
-            pageHeightPx: 0,
-        },
-        layoutMode: LayoutModes.FULLSCREEN,
-        paginationEnabled: false,
-        navigationPadEnabled: false,
-    } as const;
-
-    /**
-     * Compute the additional layout state based on the provided page and grid
-     * dimensions.
-     */
-    const layoutParametersForDimensions = (
-        pageDimensions:
-            | {
-                  pageHeightPx: never;
-                  pageWidthPx: never;
-              }
-            | {
-                  pageHeightPx: number;
-                  pageWidthPx: number;
-              },
-        gridDimensions,
-    ) => {
-        const {pageWidthPx, pageHeightPx} = pageDimensions;
-
-        // Determine the device type and orientation.
-        const deviceOrientation =
-            pageWidthPx > pageHeightPx
-                ? DeviceOrientations.LANDSCAPE
-                : DeviceOrientations.PORTRAIT;
-        const deviceType =
-            Math.min(pageWidthPx, pageHeightPx) > tabletCutoffPx
-                ? DeviceTypes.TABLET
-                : DeviceTypes.PHONE;
-
-        // Using that information, make some decisions (or assumptions)
-        // about the resulting layout.
-        const navigationPadEnabled = deviceType === DeviceTypes.TABLET;
-        const paginationEnabled =
-            deviceType === DeviceTypes.PHONE &&
-            deviceOrientation === DeviceOrientations.PORTRAIT;
-
-        const deviceInfo = {deviceOrientation, deviceType} as const;
-        const layoutOptions = {
-            navigationPadEnabled,
-            paginationEnabled,
-            // HACK(charlie): It's not great that we're making assumptions about
-            // the toolbar (which is rendered by webapp, and should always be
-            // visible and anchored to the bottom of the page for phone and
-            // tablet exercises). But this is primarily a heuristic (the goal is
-            // to preserve a 'good' amount of space between the top of the
-            // keypad and the top of the page) so we afford to have some margin
-            // of error.
-            toolbarEnabled: true,
-        } as const;
-
-        return {
-            ...computeLayoutParameters(
-                gridDimensions,
-                pageDimensions,
-                deviceInfo,
-                layoutOptions,
-            ),
-            // Pass along some of the layout information, so that other
-            // components in the heirarchy can adapt appropriately.
-            navigationPadEnabled,
-            paginationEnabled,
-        };
-    };
-
-    const layoutReducer = function (
-        state = initialLayoutState,
-        action: any,
-    ): LayoutState {
-        switch (action.type) {
-            case "ConfigureKeypad":
-                const {keypadType} = action.configuration;
-                const gridDimensions = {
-                    numRows: keypadForType[keypadType].rows,
-                    numColumns: keypadForType[keypadType].columns,
-                    numMaxVisibleRows: keypadForType[keypadType].maxVisibleRows,
-                    numPages: keypadForType[keypadType].numPages,
-                } as const;
-
-                return {
-                    ...state,
-                    ...layoutParametersForDimensions(
-                        state.pageDimensions,
-                        gridDimensions,
-                    ),
-                    gridDimensions,
-                };
-
-            case "SetPageSize":
-                const {pageWidthPx, pageHeightPx} = action;
-                const pageDimensions = {pageWidthPx, pageHeightPx} as const;
-
-                return {
-                    ...state,
-                    ...layoutParametersForDimensions(
-                        pageDimensions,
-                        state.gridDimensions,
-                    ),
-                    pageDimensions,
                 };
 
             default:
